Pairing engine test improvements
